+++ backend-java/src/main/java/com/codeworks/onboarding/domain/Item.java
@@ -1,0 +1,16 @@
+package com.codeworks.onboarding.domain;
+
+import lombok.Builder;
+import lombok.Getter;
+
+@Builder
+public class Item {
+    @Getter
+    private String name;
+
+    @Getter
+    private Integer quantity;
+
+    @Getter
+    private Double price;
+}