+++ backend-java/src/main/java/com/codeworks/onboarding/domain/ShippingFee.java
@@ -1,0 +1,15 @@
+package com.codeworks.onboarding.domain;
+
+import lombok.Builder;
+import lombok.Getter;
+
+import java.util.List;
+
+@Builder
+public class ShippingFee {
+    @Getter
+    private Long shipping;
+
+    @Getter
+    private List<Item> items;
+}